+++ app/Http/Controllers/API/ShippingController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\API;
 
+use App\Enums\Role;
 use App\Http\Controllers\Controller;
 use App\Http\Library\ApiHelpers;
 use App\Http\Requests\StoreShippingRequest;
@@ -10,6 +11,7 @@ use App\Http\Resources\ShippingCollection;
 use App\Http\Resources\ShippingResource;
 use App\Models\Shipping;
 use Illuminate\Http\JsonResponse;
+use Request;
 
 class ShippingController extends Controller
 {
@@ -17,9 +19,14 @@ class ShippingController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index(int $driverId) : JsonResponse
+    public function index(Request $request) : JsonResponse
     {
-        $shippings = Shipping::query()->where('driver_id', $driverId)->get();
+        $this->authorize('viewAny', [Shipping::class]);
+        $user = $request->user();
+        if (Role::isLogistician($user) || Role::isAdmin($user)) {
+            return $this->onSuccess(new ShippingCollection(Shipping::all()));
+        }
+        $shippings = Shipping::query()->where('driver_id', $user->id)->get();
         return $this->onSuccess(new ShippingCollection($shippings), 'Shippings retrieved successfully.');
     }
 
@@ -28,6 +35,7 @@ class ShippingController extends Controller
      */
     public function store(StoreShippingRequest $request) : JsonResponse
     {
+        $this->authorize('create', [Shipping::class]);
         $shipping = Shipping::create($request->validated());
         return $this->onSuccess($shipping, 'Shipping created successfully.');
     }
@@ -37,6 +45,7 @@ class ShippingController extends Controller
      */
     public function show(Shipping $shipping) : JsonResponse
     {
+        $this->authorize('view', [Shipping::class, $shipping]);
         return $this->onSuccess(new ShippingResource($shipping), 'Shipping retrieved successfully.');
     }
 
@@ -45,6 +54,7 @@ class ShippingController extends Controller
      */
     public function update(UpdateShippingRequest $request, Shipping $shipping) : JsonResponse
     {
+        $this->authorize('update', [Shipping::class, $shipping]);
         $shipping->update($request->validated());
         return $this->onSuccess($shipping, 'Shipping updated successfully.');
     }
@@ -54,18 +64,22 @@ class ShippingController extends Controller
      */
     public function destroy(Shipping $shipping) : JsonResponse
     {
+        $this->authorize('delete', [Shipping::class, $shipping]);
         $shipping->delete();
         return $this->onSuccess(null, 'Shipping deleted successfully.');
     }
 
     public function confirmStart(Shipping $shipping) : JsonResponse
     {
+        $this->authorize('change-start-end-state', [Shipping::class, $shipping]);
         $shipping->update(['shipped_date' => now()]);
         return $this->onSuccess(new ShippingResource($shipping), 'Shipping updated successfully.');
     }
 
     public function confirmEnd(Shipping $shipping) : JsonResponse
     {
+        $this->authorize('change-start-end-state', [Shipping::class, $shipping]);
+        //TODO: Какая-то хрень, похже посмотреть
         $shipping->update(['delivery_date' => null]);
         return $this->onSuccess(new ShippingResource($shipping), 'Shipping updated successfully.');
     }
