+++ app/Http/Controllers/API/ShippingOrderController.php
@@ -18,6 +18,7 @@ class ShippingOrderController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', ShippingOrder::class);
         return $this->onSuccess(new ShippingOrderCollection(ShippingOrder::all()), 'All shipping orders retrieved successfully.');
     }
 
@@ -26,6 +27,7 @@ class ShippingOrderController extends Controller
      */
     public function store(StoreShippingOrderRequest $request)
     {
+        $this->authorize('create', ShippingOrder::class);
         $shipping = ShippingOrder::create($request->validated());
         return $this->onSuccess(new ShippingOrderResource($shipping), 'Shipping created successfully.', 201);
     }
@@ -35,6 +37,7 @@ class ShippingOrderController extends Controller
      */
     public function show(ShippingOrder $shippingOrder)
     {
+        $this->authorize('view', $shippingOrder);
         return $this->onSuccess(new ShippingOrderResource($shippingOrder), 'Shipping retrieved successfully.');
     }
 
@@ -43,6 +46,7 @@ class ShippingOrderController extends Controller
      */
     public function update(UpdateShippingOrderRequest $request, ShippingOrder $shippingOrder)
     {
+        $this->authorize('update', $shippingOrder);
         $shippingOrder->update($request->validated());
         return $this->onSuccess(new ShippingOrderResource($shippingOrder), 'Shipping updated successfully.');
     }
@@ -52,6 +56,7 @@ class ShippingOrderController extends Controller
      */
     public function destroy(ShippingOrder $shippingOrder)
     {
+        $this->authorize('delete', $shippingOrder);
         $shippingOrder->delete();
         return $this->onSuccess(null, 'Shipping deleted successfully.');
     }
